+++ KyhTestingStartingCase/ShopAdmin/Services/IEmailService.cs
@@ -1,0 +1,10 @@
+﻿using MimeKit;
+
+namespace ShopAdmin.Services
+{
+    public interface IEmailService
+    {
+        public List<MimeMessage> CreatingEmails(List<string> listOfEmailsAdresses, List<string> listOfNames);
+        public void SendingEmails(List<MimeMessage> listOfEmails);
+    }
+}
